#37 moves query tests into separate files
