+++ test/queries/runner.js
@@ -1,0 +1,3 @@
+require('./insertUser.test');
+require('./getUser.test');
+require('./insertGroup.test');
