+++ test/runner.js
@@ -1,3 +1,3 @@
 require('./backend.test');
-require('./database.test');
 require('./lib/runner');
+require('./queries/runner');
